requirements.txt
chromadb
langchain
flask
openai
python-dotenv
pandas
sentence_transformers
streamlit
pyarrow